+++ fixtures/kitchensink-browser/jest.transform.js
@@ -1,5 +1,0 @@
-const babelOptions = {
-  presets: [['react-app', { useESModules: false }]],
-};
-
-module.exports = require('babel-jest').createTransformer(babelOptions);