+++ packages/tscomp-scripts/scripts/utils/createJestConfig.js
@@ -1,0 +1,167 @@
+// @remove-file-on-eject
+/**
+ * Copyright (c) 2015-present, Facebook, Inc.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+"use strict";
+
+const fs = require("fs");
+const chalk = require("react-dev-utils/chalk");
+const paths = require("../../config/paths");
+const common = require("./common");
+const mode = common.getMode(paths.appPackageJson);
+
+module.exports = (resolve, rootDir, isEjecting) => {
+  // Use this instead of `paths.testsSetup` to avoid putting
+  // an absolute filename into configuration after ejecting.
+  const setupTestsMatches = paths.testsSetup.match(/src[/\\]setupTests\.(.+)/);
+  const setupTestsFileExtension =
+    (setupTestsMatches && setupTestsMatches[1]) || "js";
+  const setupTestsFile = fs.existsSync(paths.testsSetup)
+    ? `<rootDir>/src/setupTests.${setupTestsFileExtension}`
+    : undefined;
+
+  const config = {
+    collectCoverageFrom: ["src/**/*.{js,jsx,ts,tsx}", "!src/**/*.d.ts"],
+
+    setupFiles: [
+      isEjecting
+        ? "react-app-polyfill/jsdom"
+        : require.resolve("react-app-polyfill/jsdom")
+    ],
+    setupFilesAfterEnv: setupTestsFile ? [setupTestsFile] : [],
+    testMatch: [
+      "<rootDir>/src/**/__tests__/**/*.{js,jsx,ts,tsx}",
+      "<rootDir>/src/**/?(*.)(spec|test).{js,jsx,ts,tsx}"
+    ],
+    testEnvironment: mode === "browser" ? "jsdom" : "node",
+    testURL: "http://localhost",
+    transform: Object.assign(
+      {
+        // [paths.useBabelOnly
+        //   ? '^.+\\.(js|jsx|ts|tsx)$'
+        //   : '^.+\\.(js|jsx)$']: isEjecting
+        //   ? '<rootDir>/node_modules/babel-jest'
+        //   : resolve('config/jest/babelTransform.js'),
+        ["^.+\\.(js|jsx|ts|tsx)$"]: isEjecting
+          ? "<rootDir>/node_modules/babel-jest"
+          : resolve("config/jest/babelTransform.js"),
+        "^.+\\.css$": resolve("config/jest/cssTransform.js"),
+        "^(?!.*\\.(js|jsx|ts|tsx|css|json)$)": resolve(
+          "config/jest/fileTransform.js"
+        )
+      }
+      // paths.useBabelOnly
+      //   ? undefined
+      //   : { '^.+\\.(ts|tsx)$': '<rootDir>/node_modules/ts-jest/dist/index.js' }
+    ),
+    transformIgnorePatterns: [
+      "[/\\\\]node_modules[/\\\\].+\\.(js|jsx|ts|tsx)$",
+      "^.+\\.module\\.(css|sass|scss)$"
+    ],
+    moduleNameMapper: {
+      "^react-native$": "react-native-web",
+      "^.+\\.module\\.(css|sass|scss)$": "identity-obj-proxy"
+    },
+    moduleFileExtensions: [...paths.moduleFileExtensions, "node"].filter(
+      ext => !ext.includes("mjs")
+    ),
+    watchPlugins: [
+      require.resolve("jest-watch-typeahead/filename"),
+      require.resolve("jest-watch-typeahead/testname")
+    ],
+    globals: {
+      "ts-jest": {
+        diagnostics: {
+          ignoreCodes: [
+            6059, // rootDir is expected to contain all source files
+            6133, // Declared but never read
+            6134, // Unused locals
+            6135, // Unused parameters
+            6192, // Unused import declaration
+            6196, // Declared but never used
+            6198, // Unused destructured element
+            6199, // Unused variables
+            6205, // Unused type parameters
+            18002, // The files list in config file is empty
+            18003 // No inputs were found in config file
+          ]
+        },
+      },
+    },
+  };
+  if (mode === "server") {
+    config.watchPathIgnorePatterns = [paths.appBuild];
+  } else if (mode === "lib") {
+    config.watchPathIgnorePatterns = [paths.appBuild, paths.appBuildCjs];
+  }
+  if (rootDir) {
+    config.rootDir = rootDir;
+  }
+  const overrides = Object.assign({}, require(paths.appPackageJson).jest);
+  const supportedKeys = [
+    "collectCoverageFrom",
+    "coverageReporters",
+    "coverageThreshold",
+    "extraGlobals",
+    "globalSetup",
+    "globalTeardown",
+    "resetMocks",
+    "resetModules",
+    "snapshotSerializers",
+    "watchPathIgnorePatterns"
+  ];
+  if (overrides) {
+    supportedKeys.forEach(key => {
+      if (overrides.hasOwnProperty(key)) {
+        config[key] = overrides[key];
+        delete overrides[key];
+      }
+    });
+    const unsupportedKeys = Object.keys(overrides);
+    if (unsupportedKeys.length) {
+      const isOverridingSetupFile =
+        unsupportedKeys.indexOf("setupFilesAfterEnv") > -1;
+
+      if (isOverridingSetupFile) {
+        console.error(
+          chalk.red(
+            "We detected " +
+              chalk.bold("setupFilesAfterEnv") +
+              " in your package.json.\n\n" +
+              "Remove it from Jest configuration, and put the initialization code in " +
+              chalk.bold("src/setupTests.js") +
+              ".\nThis file will be loaded automatically.\n"
+          )
+        );
+      } else {
+        console.error(
+          chalk.red(
+            "\nOut of the box, tscomp only supports overriding " +
+              "these Jest options:\n\n" +
+              supportedKeys
+                .map(key => chalk.bold("  \u2022 " + key))
+                .join("\n") +
+              ".\n\n" +
+              "These options in your package.json Jest configuration " +
+              "are not currently supported by tscomp:\n\n" +
+              unsupportedKeys
+                .map(key => chalk.bold("  \u2022 " + key))
+                .join("\n") +
+              "\n\nIf you wish to override other Jest options, you need to " +
+              "eject from the default setup. You can do so by running " +
+              chalk.bold("npm run eject") +
+              " but remember that this is a one-way operation. " +
+              "tscomp will follow supported options from its " +
+              "mother project Create React App.\n"
+          )
+        );
+      }
+
+      process.exit(1);
+    }
+  }
+  return config;
+};