+++ tasks/e2e-simple-lib.sh
@@ -12,18 +12,22 @@
 # Start in tasks/ even if run from root directory
 cd "$(dirname "$0")"
 
-# CLI and app temporary locations
+# App temporary location
 # http://unix.stackexchange.com/a/84980
-temp_cli_path=`mktemp -d 2>/dev/null || mktemp -d -t 'temp_cli_path'`
 temp_app_path=`mktemp -d 2>/dev/null || mktemp -d -t 'temp_app_path'`
+custom_registry_url=http://localhost:4873
+original_npm_registry_url=`npm get registry`
+original_yarn_registry_url=`yarn config get registry`
 
 function cleanup {
   echo 'Cleaning up.'
+  ps -ef | grep 'verdaccio' | grep -v grep | awk '{print $2}' | xargs kill -9
   cd "$root_path"
-  rm "$cli_path"
   # Uncomment when snapshot testing is enabled by default:
-  # rm ./packages/react-scripts/template/src/__snapshots__/App.test.js.snap
-  rm -rf "$temp_cli_path" $temp_app_path
+  # rm ./packages/tscomp-scripts/template/src/__snapshots__/App.test.js.snap
+  rm -rf "$temp_app_path"
+  npm set registry "$original_npm_registry_url"
+  yarn config set registry "$original_yarn_registry_url"
 }
 
 # Error messages are redirected to stderr
@@ -38,10 +42,6 @@ function handle_exit {
   cleanup
   echo 'Exiting without error.' 1>&2;
   exit
-}
-
-function tscomp {
-  node "$temp_cli_path"/node_modules/tscomp/bin/tscomp.js "$@"
 }
 
 # Check for the existence of one or more files.
@@ -64,73 +64,52 @@ set -x
 cd ..
 root_path=$PWD
 
-# Clear cache to avoid issues with incorrect packages being used
-if hash yarnpkg 2>/dev/null
-then
-  # AppVeyor uses an old version of yarn.
-  # Once updated to 0.24.3 or above, the workaround can be removed
-  # and replaced with `yarnpkg cache clean`
-  # Issues:
-  #    https://github.com/yarnpkg/yarn/issues/2591
-  #    https://github.com/appveyor/ci/issues/1576
-  #    https://github.com/facebook/create-react-app/pull/2400
-  # When removing workaround, you may run into
-  #    https://github.com/facebook/create-react-app/issues/2030
-  case "$(uname -s)" in
-    *CYGWIN*|MSYS*|MINGW*) yarn=yarn.cmd;;
-    *) yarn=yarnpkg;;
-  esac
-  $yarn cache clean
-fi
+# if hash npm 2>/dev/null
+# then
+#   npm i -g npm@latest
+# fi
 
-# We need to install create-react-app deps to test it
-npm install
+# Bootstrap monorepo
+yarn
 
-# If the node version is < 6, the script should just give an error.
-nodeVersion=`node --version | cut -d v -f2`
-nodeMajor=`echo $nodeVersion | cut -d. -f1`
-nodeMinor=`echo $nodeVersion | cut -d. -f2`
-if [[ nodeMajor -lt 6 ]]
-then
-  cd $temp_app_path
-  err_output=`node "$root_path"/bin/tscomp.js test-node-version 2>&1 > /dev/null || echo ''`
-  [[ $err_output =~ You\ are\ running\ Node ]] && exit 0 || exit 1
-fi
+# Start local registry
+tmp_registry_log=`mktemp`
+(cd && nohup npx verdaccio@3.8.2 -c "$root_path"/tasks/verdaccio.yaml &>$tmp_registry_log &)
+# Wait for `verdaccio` to boot
+grep -q 'http address' <(tail -f $tmp_registry_log)
 
-npm install
+# Set registry to local registry
+npm set registry "$custom_registry_url"
+yarn config set registry "$custom_registry_url"
 
-if [ "$USE_YARN" = "yes" ]
-then
-  # Install Yarn so that the test can use it to install packages.
-  npm install -g yarn
-  yarn cache clean
-fi
+# Login so we can publish packages
+(cd && npx npm-auth-to-token@1.0.0 -u user -p password -e user@example.com -r "$custom_registry_url")
+
+# Lint own code
+./node_modules/.bin/eslint --max-warnings 0 packages/create-tscomp-project/
+./node_modules/.bin/eslint --max-warnings 0 packages/tscomp-scripts/
 
 # ******************************************************************************
-# Next, pack tscomp so we can verify that it work.
+# First, test the tscomp development environment.
+# This does not affect our users but makes sure we can develop it.
 # ******************************************************************************
 
-# Pack CLI
-# cd "$root_path"/packages/create-react-app
-cli_path=$PWD/`npm pack`
+git clean -df
+./tasks/publish.sh --yes --force-publish=* --skip-git --cd-version=prerelease --exact --npm-tag=latest
 
 # ******************************************************************************
-# Now that we have packed them, create a clean app folder and install them.
+# Install tscomp-scripts prerelease via create-tscomp-project prerelease.
 # ******************************************************************************
 
-# Install the CLI in a temporary location
-cd "$temp_cli_path"
+# Install the app in a temporary location
+cd $temp_app_path
+npx create-tscomp-project lib test-library
 
-# Initialize package.json before installing the CLI because npm will not install
-# the CLI properly in the temporary location if it is missing.
-npm init --yes
-
-# Now we can install the CLI from the local package.
-npm install "$cli_path"
-
+# TODO: verify we installed prerelease
 
 # ******************************************************************************
-# Common test utils
+# Now that we used create-tscomp-project to create an app depending on tscomp-scripts,
+# let's make sure all npm scripts are in the working state.
 # ******************************************************************************
 
 function test_change_outdir {
@@ -154,22 +133,26 @@ function test_change_outdir {
   rm -rf other
 }
 
-# ******************************************************************************
-# ******************************************************************************
-# ******************************************************************************
-# Test the library template
-# ******************************************************************************
-# ******************************************************************************
-# ******************************************************************************
+function verify_module_scope {
+  # Create stub json file
+  echo "{}" >> sample.json
 
-# Install the app in a temporary location
-cd $temp_app_path
-tscomp new lib --scripts-version="$cli_path" test-library
+  # Save App.tsx, we're going to modify it
+  cp src/index.ts src/index.ts.bak
 
-# ******************************************************************************
-# Now that we used tscomp to create a library,
-# let's make sure all npm scripts are in the working state.
-# ******************************************************************************
+  # Add an out of scope import
+  echo "import sampleJson from '../sample'" | cat - src/index.ts > src/index.ts.temp && mv src/index.ts.temp src/index.ts
+
+  # Make sure the build fails
+  yarn build; test $? -eq 1 || exit 1
+  # TODO: check for error message
+
+  rm sample.json
+
+  # Restore index.ts
+  rm src/index.ts
+  mv src/index.ts.bak src/index.ts
+}
 
 # Enter the app directory
 cd test-library
@@ -179,6 +162,8 @@ yarn build
 # Check for expected output
 exists lib/index.js
 exists lib/index.d.ts
+exists cjs/index.js
+exists cjs/index.d.ts
 
 # Run tests with CI flag
 CI=true yarn test
@@ -193,19 +178,21 @@ test_change_outdir index.js index.d.ts
 # ******************************************************************************
 
 # Eject...
-echo yes | yarn eject
-
-# ...but still link to the local packages
-yarn add "$root_path"
+echo yes | npm run eject
 
 # Test the build
 yarn build
 # Check for expected output
 exists lib/index.js
 exists lib/index.d.ts
+exists cjs/index.js
+exists cjs/index.d.ts
 
-# Run tests
-CI=true yarn test
+# Run tests, overriding the watch option to disable it.
+# `CI=true yarn test` won't work here because `yarn test` becomes just `jest`.
+# We should either teach Jest to respect CI env variable, or make
+# `scripts/test.js` survive ejection (right now it doesn't).
+yarn test --watch=no
 # Uncomment when snapshot testing is enabled by default:
 # exists src/__snapshots__/App.test.js.snap