+++ tasks/publish.sh
@@ -26,14 +26,15 @@ set -x
 cd ..
 root_path=$PWD
 
-# if [ -z $CI ]; then
-#   yarn compile:lockfile
-# fi
+if [ -z $CI ]; then
+  yarn compile:lockfile
+fi
 
-# if [ -n "$(git status --porcelain)" ]; then
-#   echo "Your git status is not clean. Aborting.";
-#   exit 1;
-# fi
+if [ -n "$(git status --porcelain)" ]; then
+  git diff
+  echo "Your git status is not clean. Aborting.";
+  exit 1;
+fi
 
 # Get 2FA when not CI
 otp=""
@@ -43,4 +44,4 @@ if [ -z $CI ]; then
 fi
 
 # Go!
-NPM_CONFIG_OTP="$otp" npm publish "$@"
+NPM_CONFIG_OTP="$otp" ./node_modules/.bin/lerna publish --independent --npm-client=npm "$@"